EXP-4394: complete DSFinV-K cash point closing transaction model against the Fiskaly spec
Make the library a faithful, complete mapping of the cash point closing
transaction so the optionality matches the spec and consumers control what is sent:

- head: add optional name/user/buyer; make tx_id optional
- data: add optional notes; mark payment_types/amounts_per_vat_id/lines optional
- payment_types: add optional name/foreign_amount
- lines: add optional in_house/voucher_id/source_cash_register/references
- business_case: add optional name/purchaser_agency_id; add the 6 missing enum values
- item: add optional gtin/quantity_factor/quantity_measure/group_id/group_name,
  plus base/discounts/extra amounts (VatAmountBreakdown) and sub_items
- fix vat_definition_export_id int -> long (spec allows up to 9_999_999_999)
- first/last_transaction_export_id are now supplied by the caller (the "0"
  zero-transaction-day default moves to the consumer, out of the mapper)

+++ src/Fiskaly/Mews.Fiscalizations.Fiskaly/Models/DSFinVK/CashPointClosings/AmountPerVat.cs
@@ -1,7 +1,8 @@
 namespace Mews.Fiscalizations.Fiskaly.Models.DSFinVK.CashPointClosings;
 
+// vat_definition_export_id is an integer up to 9_999_999_999 per spec, so it must be a long.
 public sealed record AmountPerVat(
-    int VatDefinitionExportId,
+    long VatDefinitionExportId,
     decimal GrossAmount,
     decimal NetAmount,
     decimal TaxAmount
